forecast_simulator/requirements.txt
kafka-python==1.4.7
Faker==20.1.0